# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is solution iv avito backend intership 2022 by prr133f",
        "title": "Avito Intership 2022 Solution API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/balance": {
            "put": {
                "description": "Returns new balance of user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Balance"
                ],
                "summary": "Add balance to user",
                "parameters": [
                    {
                        "description": "DTO for adding user balance",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserAddBalance"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AddBalance"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/balance/{id}": {
            "get": {
                "description": "Returns balance of user by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Balance"
                ],
                "summary": "Get balance of user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uuid of user",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetBalance"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/commit": {
            "put": {
                "description": "Commites previous reserved amount",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Balance"
                ],
                "summary": "Commit reserved amount",
                "parameters": [
                    {
                        "description": "DTO for commiting amount",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserCommitAmount"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ReserveBalance"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/report/{date}": {
            "put": {
                "description": "Returns monthly margin report",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Get monthly report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "date in format YYYY-MM",
                        "name": "date",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.MonthlyReport"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/reserve": {
            "put": {
                "description": "Reserves a part of user balance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Balance"
                ],
                "summary": "Reserve part of user balance",
                "parameters": [
                    {
                        "description": "DTO for reserving part of user balance",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserReserveAmount"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ReserveBalance"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/restore": {
            "put": {
                "description": "Restores a part of reserved user balance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Balance"
                ],
                "summary": "Restore part of reserved user balance",
                "parameters": [
                    {
                        "description": "DTO for restoring part of reserved user balance",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserRestoreReservedAmount"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetBalance"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AddBalance": {
            "type": "object",
            "properties": {
                "new_balance": {
                    "type": "number"
                }
            }
        },
        "models.Error": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string"
                }
            }
        },
        "models.GetBalance": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                }
            }
        },
        "models.MonthlyReport": {
            "type": "object",
            "properties": {
                "margin": {
                    "type": "number"
                },
                "service_name": {
                    "type": "string"
                }
            }
        },
        "models.ReserveBalance": {
            "type": "object",
            "properties": {
                "reserved_balance": {
                    "type": "number"
                }
            }
        },
        "models.UserAddBalance": {
            "type": "object",
            "properties": {
                "add_balance": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "models.UserCommitAmount": {
            "type": "object",
            "properties": {
                "commited_amount": {
                    "type": "number"
                },
                "order_id": {
                    "type": "string"
                },
                "service_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.UserReserveAmount": {
            "type": "object",
            "properties": {
                "order_id": {
                    "type": "string"
                },
                "reserved_amount": {
                    "type": "number"
                },
                "service_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.UserRestoreReservedAmount": {
            "type": "object",
            "properties": {
                "order_id": {
                    "type": "string"
                },
                "restored_amount": {
                    "type": "number"
                },
                "service_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    }
}